requirements.txt
streamlit
PyPDF2
langchain-mistralai
langchain-huggingface
langchain-text-splitters
langchain-community
langchain-core
faiss-cpu